LPS-62205 Move
